refactor(aio): refine code style of aio module
