+++ src/core/aio/native_linux_aio_provider.cpp
@@ -31,9 +31,8 @@
 
 namespace dsn {
 
-native_linux_aio_provider::native_linux_aio_provider(disk_engine *disk,
-                                                     aio_provider *inner_provider)
-    : aio_provider(disk, inner_provider)
+native_linux_aio_provider::native_linux_aio_provider(disk_engine *disk)
+    : aio_provider(disk)
 {
     memset(&_ctx, 0, sizeof(_ctx));
     auto ret = io_setup(128, &_ctx); // 128 concurrent events
@@ -93,7 +92,7 @@ aio_context *native_linux_aio_provider::prepare_aio_context(aio_task *tsk)
     return new linux_disk_aio_context(tsk);
 }
 
-void native_linux_aio_provider::aio(aio_task *aio_tsk) { aio_internal(aio_tsk, true); }
+void native_linux_aio_provider::submit_aio_task(aio_task *aio_tsk) { aio_internal(aio_tsk, true); }
 
 void native_linux_aio_provider::get_event()
 {
