+++ src/core/aio/sim_aio_provider.cpp
@@ -31,14 +31,12 @@ namespace aio {
 
 DEFINE_TASK_CODE(LPC_NATIVE_AIO_REDIRECT, TASK_PRIORITY_HIGH, THREAD_POOL_DEFAULT)
 
-sim_aio_provider::sim_aio_provider(disk_engine *disk, aio_provider *inner_provider)
-    : native_linux_aio_provider(disk, inner_provider)
+sim_aio_provider::sim_aio_provider(disk_engine *disk)
+    : native_linux_aio_provider(disk)
 {
 }
 
-sim_aio_provider::~sim_aio_provider(void) {}
-
-void sim_aio_provider::aio(aio_task *aio)
+void sim_aio_provider::submit_aio_task(aio_task *aio)
 {
     error_code err;
     uint32_t bytes;
